chore: remove pfcp session create
